Refactor: Moved FolderTree component to widget folder Also, fixed styling issues in ValidateView
